% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr_uniqueness.R
\name{add_attr_uniqueness}
\alias{add_attr_uniqueness}
\title{add_attr_uniqueness}
\usage{
add_attr_uniqueness(network, qca_data, attr = NULL, bipartite = NULL)
}
\arguments{
\item{network}{An object of class 'network'.}

\item{qca_data}{The QCA-dataset you want to extend.}

\item{attr}{The node-level attribute for that you would like to calculate an uniqueness-score.
Note that this function works only with node-attributes of type 'character'!}

\item{bipartite}{If the network is bipartite, select the level the cases and the
outcome are on. If they are on the first level (sometimes referred as the
actor level) of a bipartite network, choose 'b1', if they are on the second level
(sometimes referred as the event level), choose 'b2'. Otherwise omit it.}
}
\value{
The preexisting QCA-dataset enhanced with a new column consisting of the
uniqueness scores for each node-level attribute of the selected cases.
}
\description{
add_attr_uniqueness
}
\details{
Sometimes it might be interesting to know, if an attribute of a case is rather unique (or very common)
and whether this uniqueness might be a necessary and/or sufficient condition for the outcome.
You can add a simple uniqueness-metric for a selected node-level attribute of your cases with this function.
}
\examples{
\dontrun{qca_data <- add_attr_uniqueness(network, qca_data, attr = "Type", bipartite = "b2")}

}
\references{
Butts, Carter T. (2008). network: a Package for Managing Relational Data in R. Journal of Statistical Software, 24(2).
\url{https://www.jstatsoft.org/v24/i02/paper}.
}
