% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_brokerage.R
\name{add_brokerage}
\alias{add_brokerage}
\title{add_brokerage}
\usage{
add_brokerage(
  network,
  qca_data,
  attr = NULL,
  role = NULL,
  bipartite = NULL,
  normalization = FALSE
)
}
\arguments{
\item{network}{An object of class 'network'.}

\item{qca_data}{The QCA-dataset you want to extend.}

\item{attr}{The node-level attribute that defines the group membership of the vertex and its alters
you would like to calculate the brokerage-scores for.}

\item{role}{Defines the brokerage role, you would like to calculate the score for. Choose between 'Coordinator',
'Itinerant', 'Gatekeeper', 'Representative', 'Liaison' or 'Cumulative' (see the details-section for more information).}

\item{bipartite}{If the network is bipartite, select the level the cases and the
outcome are on. If they are on the first level (sometimes referred as the
actor level) of a bipartite network, choose 'b1', if they are on the second level
(sometimes referred as the event level), choose 'b2'. Otherwise omit it.}

\item{normalization}{Set TRUE, if you would like to include normalized brokerage-scores.}
}
\value{
The preexisting QCA-dataset enhanced with a new column with the brokerage-scores of interest.
}
\description{
Performs a brokerage analysis and adds a user-defined brokerage-score for each case (node) to the QCA data-set.
}
\details{
This function relies on Butts (2020) brokerage function within his great sna package. Butts (2020),
following Gould and Fernandez (1989), implemented a bunch predefined brokerage roles.
They analyze the role a specific vertex plays between two vertex groups by calculating scores for the
brokerage role of interest. The group membership can be chosen by defining the the attr argument.
Butts (2008) describes the brokerage roles as following:

\itemize{
    \item{Coordinator role; the broker mediates contact between two individuals from his or her own group. Two-path structure: A -> A -> A}
    \item{Itinerant broker role; the broker mediates contact between two individuals from a single group to which he or she does not belong. Two-path structure: A -> B -> A}
    \item{Gatekeeper role; the broker mediates an incoming contact from an out-group member to an in-group member. Two-path structure: A -> B -> B}
    \item{Representative role; the broker mediates an outgoing contact from an in-group member to an out-group member. Two-path structure: A -> A -> B}
    \item{Liaison role; the broker mediates contact between two individuals from different groups, neither of which is the group to which he or she belongs. Two-path structure: A -> B -> C}
    \item{Total (cumulative) brokerage role occupancy. (Any of the above two-paths.)}
}

For more detail, see the description of the sna package and its brokerage-function! Please keep in mind, that the roles for bipartite networks can be
defined differently, since e.g. actors and events might have different vertex attributes and are probably not members of the same group by definition.
}
\examples{
\dontrun{qca_data <- add_brokerage(network, qca_data, attr = "Type", role = "Itinerant", bipartite = "b2", normalization = TRUE)}

}
\references{
Butts, Carter T. (2008). network: a Package for Managing Relational Data in R. Journal of Statistical Software, 24(2).
\url{https://www.jstatsoft.org/v24/i02/paper}.

Butts, Carter T. (2020). sna: Tools for Social Network Analysis. R package
version 2.6. \url{https://CRAN.R-project.org/package=sna}.

Gould, R.V. and Fernandez, R.M. (1989). Structures of Mediation: A Formal Approach to Brokerage in Transaction Networks.
Sociological Methodology, 19: 89-126.

Müller, Kirill and Wickham, Hadley (2021). tibble: Simple Data Frames. R package
version 3.1.4. \url{https://CRAN.R-project.org/package=tibble}.
}
