% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node_connectivity.R
\name{add_node_connectivity}
\alias{add_node_connectivity}
\title{add_node_connectivity}
\usage{
add_node_connectivity(network, qca_data, count = NULL, bipartite = NULL)
}
\arguments{
\item{network}{An object of class 'network'.}

\item{qca_data}{The QCA-dataset you want to extend.}

\item{count}{The connectivity-count you want to add to the QCA-dataset as a variable.
Right now, threepaths and fourcycles are available. Select them with e.g. 'count = "Fourcycles"'}

\item{bipartite}{If the network is bipartite, select the level the cases and the
outcome are on. If they are on the first level (sometimes referred as the
actor level) of a bipartite network, choose 'b1', if they are on the second level
(sometimes referred as the event level), choose 'b2'. Otherwise omit it.}
}
\value{
The preexisting QCA-dataset enhanced with a new column with the threepaths or fourcycles
the cases (or nodes) are a member of.
}
\description{
Adds counts for the connectivity of the cases (nodes).
}
\details{
Sometimes it might be interesting to know how well the the nodes that are the cases for the QCA
are connected within the network and if that this connectivity could be a condition for the outcome that
you want to examine. This function adds a new column to your QCA-dataset with the number of threepaths
and/or fourcycles that the individual cases/nodes are a member of.
}
\examples{
qca_data <- add_node_connectivity(qca_data, network, count = "Fourcycles", "b2")

}
\references{
Butts, Carter T. (2008). network: a Package for Managing Relational Data in R. Journal of Statistical Software, 24(2).
\url{https://www.jstatsoft.org/v24/i02/paper}.

Butts, Carter T. (2020). sna: Tools for Social Network Analysis. R package
version 2.6. \url{https://CRAN.R-project.org/package=sna}.
}
