% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_node_indice.R
\name{add_node_indice}
\alias{add_node_indice}
\title{add_node_indice}
\usage{
add_node_indice(network, qca_data, indice = NULL, bipartite = NULL, ...)
}
\arguments{
\item{network}{An object of class 'network'.}

\item{qca_data}{The QCA-dataset you want to extend.}

\item{indice}{The SNA node-level indice you want to add to the QCA-dataset as a variable
and its parameters (e.g. degree, cmode = "indegree"). For details, see the SNA package.}

\item{bipartite}{If the network is bipartite, select the level the cases and the
outcome are on. If they are on the first level (sometimes referred as the
actor level) of a bipartite network, choose 'b1', if they are on the second level
(sometimes referred as the event level), choose 'b2'. Otherwise omit it.}
}
\value{
The preexisting qca-dataset enhanced with a new column consisting of the
specified sna node-level indice.
}
\description{
Adds a SNA node-level indice to the QCA data-set
}
\details{
With the add_node_indice function you can add the specific node-level metrics
for each of the cases to the QCA-dataset. The 'indice' parameter is basically the command
for the SNA package to calculate the indice you like to add.

Right now, it should be compatible with betweenness, closeness, degree, stresscent,
graphcent, and evcent. For the specific commands and more options, please refer to the
reference manual of the SNA-package or the help file for the specific functions
(e.g. \code{\link[sna]{degree}}).
}
\examples{
qca_data <- add_node_indice(network, qca_data, indice = degree, cmode = "indegree", "b2")

}
\references{
Butts, Carter T. (2008). network: a Package for Managing Relational Data in R. Journal of Statistical Software, 24(2).
\url{https://www.jstatsoft.org/v24/i02/paper}.

Butts, Carter T. (2020). sna: Tools for Social Network Analysis. R package
version 2.6. \url{https://CRAN.R-project.org/package=sna}.
}
