% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qcadata.R
\name{create_qcadata}
\alias{create_qcadata}
\title{create_qcadata}
\usage{
create_qcadata(network, attr_case, attr_outcome, b2 = FALSE)
}
\arguments{
\item{network}{An object of class 'network'}

\item{attr_case}{The node-level attributes that are the cases (usually vertex.names)}

\item{attr_outcome}{The node_level attributes that are defined as the outcome}

\item{b2}{Use b2 if the cases and the outcome are on the second level (sometimes referred as the
event level) of a bipartite network. Otherwise leave empty.}
}
\value{
Creates a data set (tibble) with two columns ('Cases' and the corresponding 'Outcome')
}
\description{
Function to create a basic data set of cases and outcome.
}
\details{
Creating this data set and defining cases and outcome is always the first step.
Other variables or network-indices can be added in subsequent steps.
Please note that the tibble format is used due to convenience. After adding the desired variables
the qcadata-tibble can be finalized/converted with the finalize_qcadata-function!
}
