% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attr_uniqueness.R
\name{add_attr_uniqueness}
\alias{add_attr_uniqueness}
\title{add_attr_uniqueness}
\usage{
add_attr_uniqueness(network, qca_data, attr = NULL, bipartite = NULL)
}
\arguments{
\item{network}{An object of class 'network'.}

\item{qca_data}{The qca-dataset you want to extend.}

\item{attr}{The node-level attribute that you would like to calculate an uniqueness-
score. Note that this function works only with node-attributes of type 'character'!}

\item{bipartite}{If the network is bipartite, select the level the cases and the
outcome are on. If they are on the first level (sometimes referred as the
actor level) of a bipartite network, choose 'b1', if they are on the second level
(sometimes referred as the event level), choose 'b2'. Otherwise omit it.}
}
\description{
add_attr_uniqueness
}
\examples{
qca_data <- add_attr_uniqueness(network, qca_data, "Type", bipartite = "b2")
}
